--- proxy-plugin/src/plugin.yml
@@ -1,4 +1,0 @@
-name: Infinite-Minecraft-Players-POC-Proxy-Plugin
-main: fr.raraph84.infinite_minecraft_players_poc.proxy_plugin.MinecraftInfinitePlayersPOCProxyPlugin
-version: 1.0
-author: Raraph84
